-enh clean up the llama stuff, make it a wrapper class and move out of main
