+++ Server/src/common/common.cpp
