+++ Server/to_speak
@@ -1,1 +1,1 @@
- Het gaat goed met me en ik ben klaar om je te helpen waarmee dan ook.
+ Berlin is de hoofdstad en grootste stad van Duitsland, gelegen in Noord-Centraal-Europa. Met een bevolking van bijna 3,7 miljoen mensen is het de meest dichtbevolkte deelstaat van Duitsland.
